refactor: update video saving functionality and add VAE to ComfyUI image conversion.
