+++ lightx2v/utils/utils.py
@@ -1,14 +1,15 @@
-import os
+from typing import Optional
 from einops import rearrange
 from loguru import logger
 
 import torch
-import torchvision
 import numpy as np
-import imageio
-import random
-
 import os
+import imageio
+import torchvision
+import random
+import subprocess
+import imageio_ffmpeg as ffmpeg
 
 
 def seed_all(seed):
@@ -50,7 +51,7 @@ def save_videos_grid(videos: torch.Tensor, path: str, rescale=False, n_rows=1, f
 
 def cache_video(
     tensor,
-    save_file,
+    save_file: str,
     fps=30,
     suffix=".mp4",
     nrow=8,
@@ -73,7 +74,7 @@ def cache_video(
     for _ in range(retry):
         try:
             # preprocess
-            tensor = tensor.clamp(min(value_range), max(value_range))
+            tensor = tensor.clamp(min(value_range), max(value_range))  # type: ignore
             tensor = torch.stack(
                 [torchvision.utils.make_grid(u, nrow=nrow, normalize=normalize, value_range=value_range) for u in tensor.unbind(2)],
                 dim=1,
@@ -94,3 +95,160 @@ def cache_video(
     else:
         logger.info(f"cache_video failed, error: {error}", flush=True)
         return None
+
+
+def vae_to_comfyui_image(vae_output: torch.Tensor) -> torch.Tensor:
+    """
+    Convert VAE decoder output to ComfyUI Image format
+
+    Args:
+        vae_output: VAE decoder output tensor, typically in range [-1, 1]
+                    Shape: [B, C, T, H, W] or [B, C, H, W]
+
+    Returns:
+        ComfyUI Image tensor in range [0, 1]
+        Shape: [B, H, W, C] for single frame or [B*T, H, W, C] for video
+    """
+    # Handle video tensor (5D) vs image tensor (4D)
+    if vae_output.dim() == 5:
+        # Video tensor: [B, C, T, H, W]
+        B, C, T, H, W = vae_output.shape
+        # Reshape to [B*T, C, H, W] for processing
+        vae_output = vae_output.permute(0, 2, 1, 3, 4).reshape(B * T, C, H, W)
+
+    # Normalize from [-1, 1] to [0, 1]
+    images = (vae_output + 1) / 2
+
+    # Clamp values to [0, 1]
+    images = torch.clamp(images, 0, 1)
+
+    # Convert from [B, C, H, W] to [B, H, W, C]
+    images = images.permute(0, 2, 3, 1).cpu()
+
+    return images
+
+
+def save_to_video(
+    images: torch.Tensor,
+    output_path: str,
+    fps: float = 24.0,
+    method: str = "imageio",
+    lossless: bool = False,
+    output_pix_fmt: Optional[str] = "yuv420p",
+) -> None:
+    """
+    Save ComfyUI Image tensor to video file
+
+    Args:
+        images: ComfyUI Image tensor [N, H, W, C] in range [0, 1]
+        output_path: Path to save the video
+        fps: Frames per second
+        method: Save method - "imageio" or "ffmpeg"
+        lossless: Whether to use lossless encoding (ffmpeg method only)
+        output_pix_fmt: Pixel format for output (ffmpeg method only)
+    """
+    assert images.dim() == 4 and images.shape[-1] == 3, "Input must be [N, H, W, C] with C=3"
+
+    # Ensure output directory exists
+    os.makedirs(os.path.dirname(output_path) or ".", exist_ok=True)
+
+    if method == "imageio":
+        # Convert to uint8
+        frames = (images * 255).cpu().numpy().astype(np.uint8)
+        imageio.mimsave(output_path, frames, fps=fps)  # type: ignore
+
+    elif method == "ffmpeg":
+        # Convert to numpy and scale to [0, 255]
+        frames = (images * 255).cpu().numpy().clip(0, 255).astype(np.uint8)
+
+        # Convert RGB to BGR for OpenCV/FFmpeg
+        frames = frames[..., ::-1].copy()
+
+        N, height, width, _ = frames.shape
+
+        # Ensure even dimensions for x264
+        width += width % 2
+        height += height % 2
+
+        # Get ffmpeg executable from imageio_ffmpeg
+        ffmpeg_exe = ffmpeg.get_ffmpeg_exe()
+
+        if lossless:
+            command = [
+                ffmpeg_exe,
+                "-y",  # Overwrite output file if it exists
+                "-f",
+                "rawvideo",
+                "-s",
+                f"{int(width)}x{int(height)}",
+                "-pix_fmt",
+                "bgr24",
+                "-r",
+                f"{fps}",
+                "-loglevel",
+                "error",
+                "-threads",
+                "4",
+                "-i",
+                "-",  # Input from pipe
+                "-vcodec",
+                "libx264rgb",
+                "-crf",
+                "0",
+                "-an",  # No audio
+                output_path,
+            ]
+        else:
+            command = [
+                ffmpeg_exe,
+                "-y",  # Overwrite output file if it exists
+                "-f",
+                "rawvideo",
+                "-s",
+                f"{int(width)}x{int(height)}",
+                "-pix_fmt",
+                "bgr24",
+                "-r",
+                f"{fps}",
+                "-loglevel",
+                "error",
+                "-threads",
+                "4",
+                "-i",
+                "-",  # Input from pipe
+                "-vcodec",
+                "libx264",
+                "-pix_fmt",
+                output_pix_fmt,
+                "-an",  # No audio
+                output_path,
+            ]
+
+        # Run FFmpeg
+        process = subprocess.Popen(
+            command,
+            stdin=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+        )
+
+        if process.stdin is None:
+            raise BrokenPipeError("No stdin buffer received.")
+
+        # Write frames to FFmpeg
+        for frame in frames:
+            # Pad frame if needed
+            if frame.shape[0] < height or frame.shape[1] < width:
+                padded = np.zeros((height, width, 3), dtype=np.uint8)
+                padded[: frame.shape[0], : frame.shape[1]] = frame
+                frame = padded
+            process.stdin.write(frame.tobytes())
+
+        process.stdin.close()
+        process.wait()
+
+        if process.returncode != 0:
+            error_output = process.stderr.read().decode() if process.stderr else "Unknown error"
+            raise RuntimeError(f"FFmpeg failed with error: {error_output}")
+
+    else:
+        raise ValueError(f"Unknown save method: {method}")
